requirements.txt
streamlit-modal==0.1.2
transformers==4.39.3
python-dotenv~=1.0.1
accelerate~=0.28.0
ctransformers~=0.2.27
langchain_openai~=0.1.20
langchain_community~=0.2.16
sentence-transformers~=2.5.1
pypdf
# langchain~=0.1.16
# langchain-openai~=0.1.20
faiss-cpu~=1.8.0